requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.15
deprecation==2.1.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gotrue==2.11.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.27.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
limits==3.14.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.1.0
netifaces==0.11.0
numpy==2.2.1
oauthlib==3.2.0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
plyer==2.1.0
postgrest==0.18.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==8.3.4
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
realtime==2.0.6
requests==2.25.1
rich==13.9.4
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
ssh-import-id==5.11
storage3==0.9.0
StrEnum==0.4.15
stripe==11.4.1
supabase==2.10.0
supafunc==0.7.0
systemd-python==234
tomli==2.2.1
tqdm==4.67.1
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==1.0.0
zope.interface==5.4.0
